Revert "LPD-51094 How can this ever be null?"
This reverts commit 2d89696bbec8f746ffc8c792a83347025bcad28f.

+++ modules/apps/sharing/sharing-search-test/src/testIntegration/java/com/liferay/sharing/search/internal/test/SharingEntrySearchDLTest.java
@@ -189,10 +189,12 @@ public class SharingEntrySearchDLTest {
 			}
 		}
 		finally {
-			_userGroupLocalService.deleteUserUserGroup(
-				_groupUser.getUserId(), userGroup);
+			if (userGroup != null) {
+				_userGroupLocalService.deleteUserUserGroup(
+					_groupUser.getUserId(), userGroup);
 
-			_userGroupLocalService.deleteUserGroup(userGroup);
+				_userGroupLocalService.deleteUserGroup(userGroup);
+			}
 		}
 	}
